Initialize P2Mobile: Portal 2 in Alien Swarm Engine (Fork) with Audit State

+++ src/public/tier2/riff.h
@@ -1,4 +1,4 @@
-//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
+//========= Copyright ï¿½ 1996-2005, Valve Corporation, All rights reserved. ============//
 //
 // Purpose: 
 //
@@ -16,6 +16,8 @@
 #pragma once
 
 #include "commonmacros.h"
+#include "tier0/basetypes.h"
+#include <stdint.h>
 
 
 //-----------------------------------------------------------------------------
@@ -25,12 +27,12 @@
 class IFileReadBinary
 {
 public:
-	virtual int open( const char *pFileName ) = 0;
-	virtual int read( void *pOutput, int size, int file ) = 0;
-	virtual void close( int file ) = 0;
-	virtual void seek( int file, int pos ) = 0;
-	virtual unsigned int tell( int file ) = 0;
-	virtual unsigned int size( int file ) = 0;
+	virtual int64 open( const char *pFileName ) = 0;
+	virtual int read( void *pOutput, int size, int64 file ) = 0;
+	virtual void close( int64 file ) = 0;
+	virtual void seek( int64 file, int pos ) = 0;
+	virtual unsigned int tell( int64 file ) = 0;
+	virtual unsigned int size( int64 file ) = 0;
 };
 
 
@@ -99,11 +101,11 @@ private:
 class IFileWriteBinary
 {
 public:
-	virtual int create( const char *pFileName ) = 0;
-	virtual int write( void *pData, int size, int file ) = 0;
-	virtual void close( int file ) = 0;
-	virtual void seek( int file, int pos ) = 0;
-	virtual unsigned int tell( int file ) = 0;
+	virtual int64 create( const char *pFileName ) = 0;
+	virtual int write( void *pData, int size, int64 file ) = 0;
+	virtual void close( int64 file ) = 0;
+	virtual void seek( int64 file, int pos ) = 0;
+	virtual unsigned int tell( int64 file ) = 0;
 };
 //-----------------------------------------------------------------------------
 // Purpose: Used to write a RIFF format file
